Add file.exists check

+++ app/src/org/commcare/tasks/DeleteLogs.kt
@@ -35,7 +35,7 @@ class DeleteLogs(appContext: Context, workerParams: WorkerParameters)
         val reportsToDelete = Vector<Int>()
         for (deviceReportRecord in deviceReportRecordStorage) {
             val reportFile = File(deviceReportRecord.filePath)
-            if (reportFile.lastModified() < threeMonthsAgo.time) {
+            if (reportFile.exists() && (reportFile.lastModified() < threeMonthsAgo.time)) {
                 reportFile.delete()
                 reportsToDelete.addElement(deviceReportRecord.id)
             }
